Windows: Updating Arduino drivers
These are the Arduino drivers found in zip archive of the 1.6.12
release.
Beleaving the README the drivers are signed for Windows 8 (therefore
also for newer versions?!). However, I have not installed them
currently, just to be sure, that our installer will replace the
preinstalled drivers provided by Windows.

Contributes to CURA-2713