Update translation files
These are generated from commit a127d463bcd7e23e073b93743b44b6be67ac6617 in Cura. These are only Cura's files. The Russian translation is in and the partial Korean and Japanese translations too. The rest (German, Spanish, Finnish, French, Italian, Dutch, Portuguese and Turkish) are not yet in. However, even though the Russian translation is complete, these are not all machine-readable files for the translation since Uranium is not updated. This was just to conduct a test for the Japanese and Korean languages. We'll update again when the rest of the translations are in.

Contributes to issue CURA-3487.