Move What's New pages to cura-binary-data
The HTML files are not really binary, but it's nice to keep these together since they belong together.
The installation explicitly limits them to 5 pages, as a technological check on a social limit. Hopefully this will help constrain this in the future, since it's now an extra hurdle to add more than 5.